ci(results): typecasting corrections
`coordinate_decomposition`'s use of `npt.ArrayLike` was not justified:  not all arraylikes define a `__len__`, as `int`s and `float`s are arraylike.

The smallest typing compatible with `numpy.NDArray` that A:  defines a length, B: defines a __get_item__, and C: defines a __next__ is probably the python builtin `Sequence`.