fix(Results): fixed failing `test_qblox_result_acquisitions` test
Columns generated are the 6 of `['acquisitions_index', 'bins_index', 'i', 'q', 'amplitude', 'phase']`. Only the latter 4 contain data, but all of them are columns.
This is so that the dataframes are easier to concatenate, and keep track of the indices when mixed with objects with no-data occurrences at higher abstraction levels (ie. on `Results`).