Qhc 809 implement automatic mixers calibration for qblox modules in (#917)
This PR brings Qblox automatic mixers calibration. It has been tested in
hardware already.

When the right parameters of the QCM-RF or QRM-RF modules have been set
to the right values, the modules will trigger the mixers calibration
after setting parameters like NCO or IF, depending on the calibration
type specified in the calibration type parameters. To prevent any
unexpected misleading effect on experiments, it has been agreed with
experimentalists that this parameters are set to "off" value by default.

There are three different values in Qililab: "off", "lo" and
"lo_and_sidebands" which mirror the qblox_instruments values "off", "lo"
and "lo and sidebands".

This can be triggered in two different ways, first one is setting this
parameter to a specific value in the runcard for the RF instrument that
the calibration wants to be triggered. This will trigger every time the
right parameter has been modified in the instrument, as per Qblox
hardware implementation. The second way is to trigger it manually
through the platform by callind the "calibrate_mixers" method for a
given bus.

---------

Co-authored-by: GitHub Actions bot <actions@github.com>