QHC 960 improve the drawing for qblox (#933)
Bugs
- Bug corrected, the move commands were being read correctly once but
were not being updated after - this caused problem with loops
- A 4ns has been added when an acquire_weighed command is added to account for the extra clock cycle

Improvements
- In the plots, the name of the flux will now appear as "bus name" Flux,
similar to I and Q when hardware modulation is enabled.
- An additional argument has been added, timeout allowing the user to
stop the plotting after the specified number of ns have been plotted.
The plotting might not be the precise number of ns inputted. For
example, if the timeout is 100 ns but there is a play operation of 150
ns, the plots will display the data until 150 ns

---------

Co-authored-by: Juan José Martínez Rodríguez <133863373+jjmartinezQT@users.noreply.github.com>