Database platform implementation (#928)
The main purpose of this PR is to modify stream array to use databases
straight from platform.stream_array(). The main changes implemented
include:

- Platform: added `stream_array()` (to call platform.stream_array and
avoid to hardcode the platform as an input to write it inside the
database) and `save_measurement_results()` (to save a measurement
already finished inside the database, for cases like old data or vna
results).
- database.py: File locating all the database structure following [Oscar
definition](https://www.notion.so/qilimanjaro/Database-structure-1a87eec14c53806dafe5e40df2a0dd9f)
and following SQLAlchemy code architecture:
      - Cooldown metadata.
      - Sample metadata.
- Measurement metadata (including all the functions necessary to save
the measurement results.
- DatabaseManager, a class to operate the database, the only one in the
file which is not part of the SQLAlchemy structure but works with it).
with functions to set the sample, cooldown, measurements and results and
to modify the database (`head` and `tail`)
- StreamArray: modified version of stream array working with the
database structure, the operation is similar to the previous one but now
it saves directly to the database.

Supplementary modifications:

- Added type checks on qblox modules parameters: this was a fix for some
issues that Oscar encountered while working with the databases as the
values were not correctly set by the machine.
- Added SQLAlchemy and xarray (usefull tool for measurements) to the
dependencies as they are required for databases.
- Relocated `save_results` and `load_results` from `data_management` to
`result/result_management.py` for structure consistency. The
load_results functions has been slightly changed to take into account
different structures of data.

---------

Co-authored-by: OscarQili <oscar.pihl@qilimanjaro.tech>
Co-authored-by: Juan José Martínez Rodríguez <133863373+jjmartinezQT@users.noreply.github.com>
Co-authored-by: jjmartinezQT <juanjo.martinez@qilimanjaro.tech>