Modify the SGS100a to include operation mode setting (#957)
The R&S SGS100a driver has now the capability to change the operation
mode between normal mode and bypass mode. The default mode is the normal
mode. The allowed strings for each mode
in the settings are `normal` and `bypass`. If a different bypass mode is selected that is not any of those two, the instrument raises a warning and defaults to normal mode. If the instrument is reset
the native instrument configuration defaults to normal.