Qhc 1064 improve flattop waveform for qblox compiler efficiency (#969)
Modified smoothed square waveform class `FlatTop(amplitude, duration,
smooth_duration, buffer = 0)` which works similar to the `Square`
waveform with an additional smoothing on the edges. The only additional
parameters are the smoothing duration and the buffer time. In
`QbloxCompiler` if the duration exceeds a threshold of 100 ns the pulses
are divide into two arbitrary pulses at the beggining and the end for
the smooth parts and a loop of square pulses in the middle, with the
exact same behaviour as `Square` pulses.

The previous version of this function was incomplete as the smooth was
not fully implemented leaving sharp edges