qhc-1115-bug-qblox-wait-from-65532-to-65535-raises-an-error (#1006)
Fixed a bug in the QBlox Compiler handling of the wait, long waits that
were a multiple of 65532 (the maximum wait) up to 65535 were giving out
an error. This has been solved by checking if the remainder would be
below 4. If the remainder is 0 it appends a wait of 65532 and if the
remainder is between 1 and 3, the duration of the last wait is computed
as : `(INST_MAX_WAIT + remainder) - INST_MIN_WAIT` (where
`INST_MAX_WAIT` is 65532 and `INST_MIN_WAIT` is 4) and a wait of
`INST_MIN_WAIT` is added.