[QHC-1056] Changing `desynch_sequencers` loop, from `instruments` to `instrument_controllers` (#964)
### SOLUTIONS:

The good solution would be to not do instrument.desync_sequencers() to
the instruments not present in the buses, instrument_controllers or
program, instead than desyncing all the instruments (the not connected
to, since we only connect the ones present in instrument_controllers)

So in summary we should change the lines:

```python
for instrument in self.instruments.elements:
    if isinstance(instrument, QbloxModule):
        instrument.desync_sequencers()
```

for looping the instrument_controllers one, something like:

```python
for instrument_controller in self.instrument_controllers.elements:
    for instrument in instrument_controller.modules:
        if isinstance(instrument, QbloxModule):     
            instrument.desync_sequencers()
```

Or we could check which instruments are used in the program/buses (with
a bit more complicated logic), and loop those instead.

So @juanjo.martinez if you think one solution is better than the others,
I'll implement that one. Benefits I see with each one:

- intrument_controllers: it's what the `connect` loops, so this error
would 100% not happen, since device will always exist (its created
there). (but if some instrument is defined there, and not used in the
program, we would desync it too, I don't know if that matters or not).

- program: related with the above, this one would only do desync to the
actual used instrument for that circuit. But the error could appear if
the runcard is illdefined and that instrument has not been connected.

- bus: I guess a this has only both the drawbacks, while no benefit?


#### (for now, I went with the first one, since ir was the easiest)


---

Other silly possible solutions:

- Doing the connection, to not only the instruments defined in
instruments_controllers, but all, so they have a self.device
- Appart than leave the defined instruments, also leave them defined in
instrument controllers too?

Already found silly solution:

- Comment/Remove redundant instruments in Runcard:
(We could add a specific error, pointing to this, so easier to recogn