Qhc 992 QDACII Qprogram compiler and qdac drivers for voltage lists (#968)
Implementation of the triggered pulses. The drivers for the QDAC
triggering and synchronization have been implemented in instruments and
for the Qblox. Added wait_trigger and set_trigger inside the qprogram
(including syncronization with other buses for Qblox). and added
Linspace loop for qprogram

Added Qdac compiler for Qprogram with all the necessary implementations
for active reset.