qhc-1192-implement-measureresetcalibrated-in-qprogram (#1042)
`MeasureResetCalibrated` has been implemented to enable the use of
active reset with a calibration file.
After retrieving the waveforms and weights from the calibration file,
the measure reset can be called with:
**`qprogram.qblox.measure_reset(bus='readout_bus', waveform='Measure',
weights='Weights', control_bus='drive_bus', reset_pulse='Drag')`**.
Unlike other methods, this one does not allow a mix of calibrated and
non-calibrated parameters is not allowed. This method requires the
calibration file to be used consistently across `waveform`, `weight`,
and `reset_pulse`; either for all three or for none. An error is raised
if this condition is not met.