[QHC-1209] Fix stream array get debug with calibration class (#1043)
There was a bug for users using StreamArray with calibration Qprograms.
In case a user was submitting a qprogram using `Calibration` waveforms
or blocks and bus mapping, the function `_getdebug()` inside
`StreamArray` crashed because this function compiles the qprogram, but
without access to the bus map.
This function (`_getdebug()`) is a quick debug function requested by the
measurement team that uses `QbloxCompiler().compile` to return the
Q1ASM. Therefore, it does not require access to the calibration file.
To fix the issue, the `StreamArray` now raises a traceable error, and
the bus_mapping has been added as a variable